+++ .vscode/build.sh
@@ -8,9 +8,6 @@ pushd $script_dir
 configuration=${1:-Debug}
 
 # build dll
-echo "Building for RimWorld 1.5"
-rm -f $mod_dir/1.5/Assemblies/BetterAutocastVPE.dll
-dotnet build $script_dir/mod.csproj -c ${configuration} -p:GAME_VERSION=v1.5
 echo "Building for RimWorld 1.6"
 rm -f $mod_dir/1.6/Assemblies/BetterAutocastVPE.dll
 dotnet build $script_dir/mod.csproj -c ${configuration} -p:GAME_VERSION=v1.6