+++ Source/Helpers/EnchantHelper.cs
@@ -4,11 +4,7 @@ using System.Linq;
 using HarmonyLib;
 using RimWorld;
 using Verse;
-#if v1_5
-using Ability = VFECore.Abilities.Ability;
-#else
 using Ability = VEF.Abilities.Ability;
-#endif
 
 namespace BetterAutocastVPE.Helpers;