+++ Source/Patches/CanAutoCast_EnableHandledPsycasts.cs
@@ -1,9 +1,5 @@
 ﻿using System;
-#if v1_5
-using VFECore.Abilities;
-#else
 using VEF.Abilities;
-#endif
 
 namespace BetterAutocastVPE.Patches;