+++ Source/Patches/Pawn_TryGetAttackVerb_Postfix.cs
@@ -4,11 +4,7 @@ using System.Linq;
 using BetterAutocastVPE.Helpers;
 using HarmonyLib;
 using Verse;
-#if v1_5
-using VFECore.Abilities;
-#else
 using VEF.Abilities;
-#endif
 
 namespace BetterAutocastVPE.Patches;