+++ Source/Settings/Windows/IgnoredMentalStateListWindow.cs
@@ -4,11 +4,7 @@ using System.Linq;
 using System.Runtime.CompilerServices;
 using UnityEngine;
 using Verse;
-#if v1_5
-using VFECore.UItils;
-#else
 using VEF.Utils;
-#endif
 
 namespace BetterAutocastVPE.Settings.Windows;